LPS-190886 Create test to ensure that all existing caches can be reconfigured. One important thing must be taken into account:
1.The number of the caches has to be big enough (that's why all of them are taken so that the issue is statistically reproducible in a consistent way each time).

The reason why it cannot be ensured its reproduction otherwise is because net.sf.ehcache.config.Configuration.getCacheConfigurations() is a ConcurrentHashMap, so cacheConfigurations.values() order cannot be assumed, and order is importan to reproduce the issue.

The issue is reproducible because at least one cache is removed in the cache manager and not recreated so it cannot be found in the MBeanServer.