LPD-27658 Fix current remove duplicates. It removes more than needed since it does not only remove common records between old and new package, but also existing records in new package that are not present in the old one.
# breaking

## What portal-impl/src/com/liferay/portal/upgrade/v7_0_0/UpgradeKernelPackage.java

Substitute upgradeTable last preventDuplicates parameter for uniqueColumns.

## Why

A parameter with the columns that define a row as unique need to be added to define what needs to be removed and what doesn't.

## Alternatives

Use UpgradeKernelPackage#upgradeTable(String, String, String[][], WildcardMode) when no removal of duplicates needs to be performed and use upgradeTable(String, String, String[][], WildcardMode, String[]) providing as last parameter the columns that define the record as unique to remove duplicates.

----